requirements.txt
google-cloud-speech ~= 1.3.2
gsutil ~= 4.50
